bringing it all together

+++ database/migrations/2022_07_04_181221_create_apartments_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->foreignId('caretaker_id')->constrained('users')->onDelete('cascade');
-            $table->integer('flat_rate_limit')->default(0);
-            $table->decimal('flat_rate')->default(0);
-            $table->decimal('unit_rate')->default(0);
+            $table->decimal('flat_rate_limit', 6, 2)->default(0.00);
+            $table->decimal('flat_rate', 6, 2)->default(0.00);
+            $table->decimal('unit_rate', 6, 2)->default(0.00);
             $table->timestamps();
 
             $table->index(['name', 'caretaker_id']);
